Philip Wyett eliminated another MSVC compiler warnings
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13509 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/spr3d/object/spr3d.cpp
@@ -1530,7 +1530,7 @@ bool csSprite3DMeshObject::DrawTest (iRenderView* rview, iMovable* movable)
     level_of_detail /= MAX (wor_sq_dist, SMALL_EPSILON);
 
     // reduce LOD based on field-of-view
-    float aspect = 2 * tan (camera->GetFOVAngle () * PI / 360);
+    float aspect = 2 * (float) tan (camera->GetFOVAngle () * PI / 360);
     level_of_detail *= aspect;
   }
